Fix problem button not working for invasion
Formations are updated in tick 5, invasion starts at tick 0, so those few ticks in between the formations are not up to date and thus the problem offset could not be determined.
Work around this by falling back to the invasion point, but only for the first month after the invasion starts.