Fix warning about multiple libraries found in Android
Warning: "More than one file was found with OS independent path 'lib/x86/libSDL2.so'."
This is caused by add_library(): it copies the referenced library to the output folder, so gradle found two copies of sdl2 and sdl2_mixer: one in the SDL2 module, one in the julius module.
Use link_directories() instead so it doesn't copy the .so files.