Make roads under aqueducts visible in overlays (#399)
Places where a road goes under an aqueduct are replaced with grass when
a non-water overlay is shown. This causes a "break" in the road shown.

This change detects this case, and renders the corresponding road tile.
(It also, in theory, would handle the case where a road and a wall
intersected, but I don't think this can happen.)

Note that this bug does exist in vanilla C3.