Fix corruption when earthquake destroys warehouse
An earthquake destroying a warehouse (or another building which consists of multiple parts) will result in 9 burning ruin "buildings" that are still chained together. The building on the earthquake crack is then removed, but is still linked to by the other buildings. When the earthquake hits one of those other burning ruins, the building chain is destroyed again, even though the previously deleted ruin's id might now belong to a newly built building, corrupting the tile and building.
This is now fixed by unlinking the chain when the building is destroyed.

Fixes #616