Improve number separators (#520)
* Improve number drawing with optional separators between digits

Rewrite text_draw_number to handle number formatting logic. Change the method signature to more closely match text_draw() for parity. Update all usages.
Add a configuration option to the config window to show separators or not. If turned on, digits are separated by 3 pixels. If off, the old behavior is preserved.

TODO:
Need to add translations for all languages

* Remove unneeded include

Co-authored-by: José Cadete <crudelios@gmail.com>