Change the way city health is calculated
Each house has a health rating, which can go from 0 to 100

Health rating calculation:

- +1 point for each housing level, up to +10 points
- Hospital and clinic access: +50 points
- Hospital access only: +40 points
- Clinic access only: +30 points
- Access to baths: +20 points
- Access to barber: +10 points
- +15 points for each variety of food eaten
- If the sum is higher than 100, it's capped at 100. However, if the house type requires food and it doesn't have food, it's capped at 40

The house's health rating is then used as a percentage of the total population of the house, where we obtain the number of healthy residents.

When all the houses are calculated, the final health value is the percentage between all healthy residents and total city population.